isolate functions on the edge in their own file.
also clean imports
