+++ backend/functions/src/emails.ts
@@ -13,12 +13,11 @@ import { formatNumericProbability } from 'common/pseudo-numeric'
 
 import { sendTemplateEmail, sendTextEmail } from './send-email'
 import { contractUrl, getUser, log } from 'shared/utils'
-import { buildCardUrl, getOpenGraphProps } from 'common/contract-details'
+import { getOpenGraphProps } from 'common/contract-details'
 import { notification_reason_types } from 'common/notification'
 import { Dictionary } from 'lodash'
 import { getNotificationDestinationsForUser } from 'common/user-notification-preferences'
 import { buildOgUrl } from 'common/util/og'
-
 
 export type PerContractInvestmentsData = {
   questionTitle: string
