+++ common/src/util/og.ts
@@ -20,16 +20,3 @@ export function pointsToBase64(points: { x: number; y: number }[]) {
   const floats = new Float32Array(points.flatMap(({ x, y }) => [x, y]))
   return Buffer.from(floats.buffer).toString('base64url')
 }
-
-export function base64toPoints(base64urlString: string) {
-  const b64 = base64urlString.replace(/-/g, '+').replace(/_/g, '/')
-  const bin = atob(b64)
-  const u = Uint8Array.from(bin, (c) => c.charCodeAt(0))
-  const f = new Float32Array(u.buffer)
-
-  const points = [] as { x: number; y: number }[]
-  for (let i = 0; i < f.length; i += 2) {
-    points.push({ x: f[i], y: f[i + 1] })
-  }
-  return points
-}
